PWN-4717 - fix deviceShareDeviceName and fix metadata on createWallet
