+++ app/src/main/java/org/p2p/wallet/auth/interactor/CreateWalletInteractor.kt
@@ -1,11 +1,14 @@
 package org.p2p.wallet.auth.interactor
 
+import org.p2p.solanaj.core.Account
 import org.p2p.wallet.auth.gateway.repository.GatewayServiceRepository
 import org.p2p.wallet.auth.model.PhoneNumber
 import org.p2p.wallet.auth.repository.SignUpFlowDataLocalRepository
 import org.p2p.wallet.auth.repository.UserSignUpDetailsStorage
 import org.p2p.wallet.auth.ui.smsinput.SmsInputTimer
 import org.p2p.wallet.infrastructure.network.provider.TokenKeyProvider
+import org.p2p.wallet.utils.toBase58Instance
+import timber.log.Timber
 
 class CreateWalletInteractor(
     private val gatewayServiceRepository: GatewayServiceRepository,
@@ -91,6 +94,12 @@ class CreateWalletInteractor(
             socialShareOwnerId = socialShareOwnerId
         )
 
+        tryLoadAndSaveOnboardingMetadata(
+            userAccount = signUpFlowDataRepository.userAccount,
+            mnemonicPhraseWords = userSeedPhrase,
+            ethereumPublicKey = etheriumPublicKey
+        )
+
         finishAuthFlow()
     }
 
@@ -105,5 +114,24 @@ class CreateWalletInteractor(
 
     fun setIsCreateWalletRequestSent(isSent: Boolean) {
         signUpFlowDataRepository.isCreateWalletRequestSent = isSent
+    }
+
+    private suspend fun tryLoadAndSaveOnboardingMetadata(
+        userAccount: Account?,
+        mnemonicPhraseWords: List<String>,
+        ethereumPublicKey: String
+    ) {
+        try {
+            requireNotNull(userAccount) { "loadAndSaveOnboarding: User account can't be null" }
+            require(mnemonicPhraseWords.isNotEmpty()) { "loadAndSaveOnboarding: seed phrase can't be null or empty" }
+            gatewayServiceRepository.loadAndSaveOnboardingMetadata(
+                solanaPublicKey = userAccount.publicKey.toBase58Instance(),
+                solanaPrivateKey = userAccount.keypair.toBase58Instance(),
+                userSeedPhrase = mnemonicPhraseWords,
+                etheriumAddress = ethereumPublicKey
+            )
+        } catch (error: Throwable) {
+            Timber.e(GetOnboardingMetadataFailed(error))
+        }
     }
 }
